auth restriction

+++ app/Http/Controllers/QuestionController.php
@@ -70,7 +70,11 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort(403, "Access denied");
+        }
         return view("questions.edit",compact('question'));
+       
     }
 
     /**
@@ -82,6 +86,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort(403, "Access denied");
+        }
         $question->update($request->only('title','body'));
 
         return redirect()->route('questions.index')->with('success', "Your question has been updated");
@@ -97,6 +104,9 @@ class QuestionController extends Controller
     public function destroy(Question $question)
     {
         //
+        if(\Gate::denies('delete-question',$question)){
+            abort(403, "Access denied");
+        }
         $question->delete();
 
         return redirect()->route('questions.index')->with('success',"Your question has been delete");
